add sdl, service; refactor to use services
