+++ api/src/functions/createCheckoutSession/createCheckoutSession.js
@@ -1,5 +1,5 @@
 import { logger } from 'src/lib/logger'
-import { stripe } from 'src/lib/stripe'
+import { createCheckoutSession } from 'src/services/checkouts'
 
 /**
  * The handler function is your code that processes http request events.
@@ -18,51 +18,11 @@ import { stripe } from 'src/lib/stripe'
  * function, and execution environment.
  */
 
-const getCartItems = {
-  subscription: [
-    {
-      price: 'price_1JdarsFnEy6nnTnpxOI5D4fA',
-      quantity: 1,
-    },
-  ],
-  payment: [
-    {
-      price_data: {
-        product_data: {
-          name: 'Bag of Apples',
-        },
-        unit_amount: 3000,
-        currency: 'zar',
-      },
-      quantity: 1,
-    },
-    {
-      price_data: {
-        product_data: {
-          name: 'Bunch of Bananas',
-        },
-        unit_amount: 2800,
-        currency: 'zar',
-      },
-      quantity: 1,
-    },
-  ],
-}
-
 export const handler = async (event, _context) => {
   logger.info('Invoked createCheckoutSession function')
   const mode = JSON.parse(event.body).mode
 
-  // Retrieve cart items from DB or wherever
-  const cartItems = getCartItems[mode]
-
-  const session = await stripe.checkout.sessions.create({
-    line_items: cartItems,
-    payment_method_types: ['card'],
-    mode: mode,
-    success_url: `http://localhost:8910/stripe-cart?success=true&sessionId={CHECKOUT_SESSION_ID}`,
-    cancel_url: `http://localhost:8910/stripe-cart?success=false`,
-  })
+  const session = await createCheckoutSession({ mode })
 
   return {
     statusCode: 200,
