+++ api/src/graphql/checkouts.sdl.js
@@ -1,0 +1,9 @@
+export const schema = gql`
+  type Session {
+    id: String!
+  }
+
+  type Mutation {
+    createCheckoutSession(mode: String!): Session! @skipAuth
+  }
+`
